// src/components/helpers/enums.ts
export enum FixedHeight {
  Fixed = 'fixed_height',
  Default = '',
}

export enum DisplayOnly {
  OnlyDisplay = 'display_only',
  Edit = 'edit',
  Default = '',
}

export enum Validness {
  Valid = 'valid',
  Invalid = 'invalid',
  NotValidated = 'notvalidated',
}

export enum Emptiness {
  Empty = 'empty',
  NotEmpty = 'not_empty',
}

export enum MsgType {
  Info = 'info',
  Warning = 'warning',
  Error = 'error',
  Success = 'success',
}

export enum Visibility {
  Hidden = 'hidden',
  Visible = 'visible',
}